+++ modules/uhsg_news/src/Plugin/Block/NewsPerDegreeProgramme.php
@@ -4,6 +4,9 @@ namespace Drupal\uhsg_news\Plugin\Block;
 
 use Drupal\Core\Block\BlockBase;
 use Drupal\Core\Cache\Cache;
+use Drupal\Core\Link;
+use Drupal\node\Entity\Node;
+use Drupal\views\Views;
 
 /**
  * Provides a 'news_per_degree_programme' block.
@@ -24,7 +27,9 @@ class NewsPerDegreeProgramme extends BlockBase {
     $query = \Drupal::entityQuery('node')
       ->condition('status', 1)
       ->condition('langcode', $lang)
-      ->condition('type', 'news');
+      ->condition('type', 'news')
+      ->sort('created', 'DESC')
+      ->range(0, 3);
     $group = $query->orConditionGroup()
       ->condition('field_news_degree_programme', NULL, 'IS NULL');
 
@@ -34,23 +39,41 @@ class NewsPerDegreeProgramme extends BlockBase {
       $group->condition('field_news_degree_programme', $tid);
     }
 
+    // get link to news view
+    $view = Views::getView('news');
+    $view->setDisplay('page_1');
+    $url = $view->getUrl();
+    $url->setOptions([
+      'attributes' => [
+        'class' => [
+          'box-subtitle__link',
+          'button--action',
+          'icon--arrow-right',
+          'theme-transparent',
+          'is-center-mobile'
+        ],
+      ],
+    ]);
+    $link = Link::fromTextAndUrl($this->t('More current topics'), $url)->toString();
+
     $nids = $query->condition($group)->execute();
 
     if ($nids) {
-      $nodes = \Drupal\node\Entity\Node::loadMultiple($nids);
+      $nodes = Node::loadMultiple($nids);
       $render_controller = \Drupal::entityTypeManager()->getViewBuilder('node');
       $render_output = $render_controller->viewMultiple($nodes, 'teaser');
 
-      return array(
+      return [
         '#attributes' => [
-          'class' => ['grid-container'],
+          'class' => ['clearfix', 'tube'],
         ],
         '#cache' => [
           'tags' => ['node_list'],
           'contexts' => ['active_degree_programme'],
         ],
         'content' => $render_output,
-      );
+        '#suffix' => $link
+      ];
     }
   }